Slight JS refactor, enable math in notes, choose tag color, small PHP refactor
This is mainly a first step toward a refactoring of the JS of DjLu. For now,
it's mostly a small reorganization of the code in the djlu.js file to make it
more readable, but also extracting anonymous functions into named function so
that it looks a bit cleaner regarding what's doing what.
Some parts of the code has also been cleaned up, particularly regarding handling
of modals that was unreasonably complicated.

Notes can also now contain LaTeX math between $'s that will be rendered using
MathJax.

A fixed color can be assigned to any tag from the left panel, that is now
animated.

Some small reorganizations have been done in the PHP code.