Add Google Drive storage handling for PDFs
For each paper, the user can store the PDF of the paper in its Google Drive
account.

From the UI, the user can ask the app to:
- Try to discover the PDF on the user's Google Drive (if the user added it
  himself)
- Download the PDF from the current URL and save it to Drive for the paper
- Save a file that the user uploads to the app to Drive for the paper

The user should set the setting "googleDriveRoot" in its preferences JSON file.
It is the folder DjLu will consider as root and where it will operate.