Update SSH key procedure for git clone, 1 key per repo
Because GitHub prevents users from using the same key for multiple repos, and
since there will surely be more than 1 user using GitHub, we need to be able to
given different public keys to users.

For this purpose, we use ssh config file to create "fake" host for each user,
that will redirect to the right host, and use the right SSH key.

SSH keys can be generated whenever a users ask to it from register page. They
are kept as temp keys until the user actually registers with the id of the temp
key.