Makes git optional, git clone done after register, adds settings modal
User account creation works without git because probably not everyone wants to
handle a git repo. It also allows the admin of the server to choose who can use
git and who can't, because allowing git means the user can send any file he
wants to the server.

This requires some kind of refactoring of the register / git clone process and
was also the occasion to do some cleanup in the user class for cleaner user data
handling.