Refactor tags handling
Tags are now all stored in user preferences and always get a color assigned.
This allows not to worry about the shift of automatic colors when adding new
papers in the library.

This also allows a more optimized loading of the existing tags in the user
library.

This was also the occasion to add papers count next to tags, to allow to pin /
unpin tags if you want to keep them in your library or not when they are not
used in any paper, and to fix a small bug in the JS.